Move MSAL Tests, bring back some tests too
